+++ tests/unit/test_media_player_vlc.py
@@ -24,7 +24,10 @@ from dakara_player.media_player.base import (
 )
 from dakara_player.media_player.vlc import (
     MediaPlayerVlc,
+    UnavailableInstanceError,
+    UnexpectedInstanceParameterError,
     VlcTooOldError,
+    get_instance,
     get_metadata,
     set_metadata,
 )
@@ -1422,3 +1425,29 @@ class OnPlayingThisTestCase(BaseTestCase):
 
         with self.get_instance() as (player, _, _):
             self.assertEqual(function_decorated(player), 42)
+
+
+@patch("dakara_player.media_player.vlc.vlc", autospec=True)
+class GetInstanceTestCase(TestCase):
+
+    def test_parameters_unexpected(self, mocked_vlc):
+        """Test to pass unexpected parameters."""
+        mocked_vlc.Instance.return_value = None
+        with self.assertRaises(UnexpectedInstanceParameterError):
+            get_instance(["parameter"])
+
+    def test_parameters(self, mocked_vlc):
+        """Test to pass parameters."""
+        self.assertIs(mocked_vlc.Instance.return_value, get_instance(["parameter"]))
+        mocked_vlc.Instance.assert_called_with(["parameter"])
+
+    def test_no_parameters_unavailable(self, mocked_vlc):
+        """Test unavailable instance."""
+        mocked_vlc.Instance.return_value = None
+        with self.assertRaises(UnavailableInstanceError):
+            get_instance()
+
+    def test_no_parameters(self, mocked_vlc):
+        """Test instance."""
+        self.assertIs(mocked_vlc.Instance.return_value, get_instance())
+        mocked_vlc.Instance.assert_called_with()